+++ src/main/java/com/example/models/CommentJSON.java
@@ -4,11 +4,11 @@ import java.sql.Date;
 
 public class CommentJSON {
 
-	private int id;
+	private long id;
 	private Date date;
 	private String commentaire;
-	private int id_user;
-	private int id_post;
+	private long id_user;
+	private long id_post;
 	private Post post;
 	
 	public Post getPost() {
@@ -17,10 +17,10 @@ public class CommentJSON {
 	public void setPost(Post post) {
 		this.post = post;
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	
@@ -36,16 +36,16 @@ public class CommentJSON {
 	public void setCommentaire(String commentaire) {
 		this.commentaire = commentaire;
 	}
-	public int getId_user() {
+	public long getId_user() {
 		return id_user;
 	}
-	public void setId_user(int id_user) {
+	public void setId_user(long id_user) {
 		this.id_user = id_user;
 	}
-	public int getId_post() {
+	public long getId_post() {
 		return id_post;
 	}
-	public void setId_post(int id_post) {
+	public void setId_post(long id_post) {
 		this.id_post = id_post;
 	}
