+++ src/main/java/com/example/models/PostJSON.java
@@ -5,11 +5,11 @@ import java.util.List;
 
 public class PostJSON {
 
-	private int id;
+	private long id;
 	private String titre;
 	private Date date;
 	private String contenu;
-	private int id_user;
+	private long id_user;
 	private List<Comment> comments;
 	
 	public List<Comment> getComments() {
@@ -18,10 +18,10 @@ public class PostJSON {
 	public void setComments(List<Comment> comments) {
 		this.comments = comments;
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getTitre() {
@@ -42,10 +42,10 @@ public class PostJSON {
 	public void setContenu(String contenu) {
 		this.contenu = contenu;
 	}
-	public int getId_user() {
+	public long getId_user() {
 		return id_user;
 	}
-	public void setId_user(int id_user) {
+	public void setId_user(long id_user) {
 		this.id_user = id_user;
 	}
 }
